requirements.txt
google-cloud-bigquery==3.17.2
jinja2==3.1.3
PyYAML==6.0.1